Move FlagParser into its own file
